Move statefile into state.file module
