+++ nixops/deployment.py
@@ -12,7 +12,6 @@ import exceptions
 import errno
 from collections import defaultdict
 from xml.etree import ElementTree
-import nixops.statefile
 import nixops.backends
 import nixops.logger
 import nixops.parallel
