+++ scripts/nixops
@@ -5,7 +5,7 @@ from nixops import deployment
 from nixops.nix_expr import py2nix
 from nixops.parallel import MultipleExceptions, run_tasks
 
-import nixops.statefile
+import nixops.state.file
 import prettytable
 import argparse
 import os
@@ -48,14 +48,14 @@ def sort_deployments(depls):
 # $NIXOPS_DEPLOYMENT.
 def one_or_all():
     if args.all:
-        sf = nixops.statefile.StateFile(args.state_file)
+        sf = nixops.state.file.StateFile(args.state_file)
         return sf.get_all_deployments()
     else:
         return [open_deployment()]
 
 
 def op_list_deployments():
-    sf = nixops.statefile.StateFile(args.state_file)
+    sf = nixops.state.file.StateFile(args.state_file)
     tbl = create_table([("UUID", 'l'), ("Name", 'l'), ("Description", 'l'), ("# Machines", 'r'), ("Type", 'c')])
     for depl in sort_deployments(sf.get_all_deployments()):
         tbl.add_row(
@@ -67,7 +67,7 @@ def op_list_deployments():
 
 
 def open_deployment():
-    sf = nixops.statefile.StateFile(args.state_file)
+    sf = nixops.state.file.StateFile(args.state_file)
     depl = sf.open_deployment(uuid=args.deployment)
 
     depl.extra_nix_path = sum(args.nix_path or [], [])
@@ -101,7 +101,7 @@ def modify_deployment(depl):
 
 
 def op_create():
-    sf = nixops.statefile.StateFile(args.state_file)
+    sf = nixops.state.file.StateFile(args.state_file)
     depl = sf.create_deployment()
     sys.stderr.write("created deployment ‘{0}’\n".format(depl.uuid))
     modify_deployment(depl)
@@ -189,7 +189,7 @@ def op_info():
                     ])
 
     if args.all:
-        sf = nixops.statefile.StateFile(args.state_file)
+        sf = nixops.state.file.StateFile(args.state_file)
         if not args.plain:
             tbl = create_table([('Deployment', 'l')] + table_headers)
         for depl in sort_deployments(sf.get_all_deployments()):
@@ -491,7 +491,7 @@ def op_export():
 
 
 def op_import():
-    sf = nixops.statefile.StateFile(args.state_file)
+    sf = nixops.state.file.StateFile(args.state_file)
     existing = set(sf.query_deployments())
 
     dump = json.loads(sys.stdin.read())
@@ -695,7 +695,7 @@ subparsers = parser.add_subparsers(help='sub-command help')
 def add_subparser(name, help):
     subparser = subparsers.add_parser(name, help=help)
     subparser.add_argument('--state', '-s', dest='state_file', metavar='FILE',
-                           default=nixops.statefile.get_default_state_file(), help='path to state file')
+                           default=nixops.state.file.get_default_state_file(), help='path to state file')
     subparser.add_argument('--deployment', '-d', dest='deployment', metavar='UUID_OR_NAME',
                            default=os.environ.get("NIXOPS_DEPLOYMENT", os.environ.get("CHARON_DEPLOYMENT", None)), help='UUID or symbolic name of the deployment')
     subparser.add_argument('--debug', action='store_true', help='enable debug output')
