+++ tests/__init__.py
@@ -3,14 +3,14 @@ import os
 import sys
 import threading
 from os import path
-import nixops.statefile
+import nixops.state.file
 
 _multiprocess_shared_ = True
 
 db_file = '%s/test.nixops' % (path.dirname(__file__))
 
 def setup():
-    nixops.statefile.StateFile(db_file).close()
+    nixops.state.file.StateFile(db_file).close()
 
 def destroy(sf, uuid):
     depl = sf.open_deployment(uuid)
@@ -27,7 +27,7 @@ def destroy(sf, uuid):
     depl.logger.log("deployment ‘{0}’ destroyed".format(uuid))
 
 def teardown():
-    sf = nixops.statefile.StateFile(db_file)
+    sf = nixops.state.file.StateFile(db_file)
     uuids = sf.query_deployments()
     threads = []
     for uuid in uuids:
