+++ tests/functional/__init__.py
@@ -1,6 +1,6 @@
 import os
 from os import path
-import nixops.statefile
+import nixops.state.file
 from tests import db_file
 
 
@@ -8,7 +8,7 @@ class DatabaseUsingTest(object):
     _multiprocess_can_split_ = True
 
     def setup(self):
-        self.sf = nixops.statefile.StateFile(db_file)
+        self.sf = nixops.state.file.StateFile(db_file)
 
     def teardown(self):
         self.sf.close()
